Eliminated last of the temporary code from init
I removed a load of code that was only there for testing purposes, which
have certainly expired in their usefulness. At the moment, there is some
more temp code to work out the kinks in the player/npc structure, which
will be removed when I properly get the creation code working.

That will be the next large chunk of code, aside from col_det() fixing.
Alright, peace.