Overhaul of map system
Old version only supported image ids 0-9, so of course that wasn't going
to work in the long run. So before things actually got far in this
project, I decided to remedy this. The new system uses delimiters (- as
the initiator and + as the terminator) where the id is in between the
initiator and the terminator.