Move towards multimap
Started implementing a multiple-map system, for the levels and such.
Also added the base entity class, along with config files for tile
functions and entity declarations. Each tile now can have a function
assigned to it (default being no action) which will affect the engine in
a certain way.

The move_north etc. functions are still non-functional as of yet, as I
am fleshing out how the character will move between maps, and I've yet
to figure out how exactly I want this to be done.