Multimap and Entities
This update brings a more or less fully fleshed out version of the
multimap system. The engine will now load multiple maps, and keep them
in memory within the current loaded level. This makes for quick
switching between maps. The switching is done mainly using the collision
detection system in conjuction with the addition of metadata for the
tiles and new additions to the maps themselves. The maps contain
pointers to the maps corresponding to the movement actions specified in
the tiles.

The new map additions consist of an entity list (with the entities being
referred to by number), a total of entities (located under the data
section), and the addition of n, e, s, and w variables. All of these
variables contain an integer which corresponds to the map id which is
intended to be the destination for movement.

Other than this, I've also worked more on the entity system, with the
system more or less done at this point. The only thing left to do is
implement the "facing" system (where entities can be rotated) and also
implement some sort of physics system for the physics_prop category.

Overall, a productive day waiting in the hospital.